Add git-treehash
It compares the SHA1 hashes of the *trees* of the refs named on the
command line.  This is useful, for example, if you are squashing or
otherwise cleaning up commit history, and you want to know if the
resulting tree is identical to the one you had before.

For this reason, the default refs are HEAD and ORIG_HEAD, for
comparing the trees before and after using git-rebase.